+++ src/hb/ot_shape_complex_arabic.rs
@@ -1,5 +1,6 @@
 use alloc::boxed::Box;
 
+use super::algs::*;
 use super::buffer::*;
 use super::ot_layout::*;
 use super::ot_map::*;
@@ -667,6 +668,6 @@ pub const ARABIC_SHAPER: hb_ot_complex_shaper_t = hb_ot_complex_shaper_t {
     setup_masks: Some(setup_masks_arabic_plan),
     gpos_tag: None,
     reorder_marks: Some(reorder_marks_arabic),
-    zero_width_marks: Some(ZeroWidthMarksMode::ByGdefLate),
+    zero_width_marks: HB_OT_SHAPE_ZERO_WIDTH_MARKS_BY_GDEF_LATE,
     fallback_position: true,
 };
